+++ Demo/Shared/ContentView.swift
@@ -69,10 +69,7 @@ private extension ContentView {
         } else {
             return AnyView(StackedDeck(
                 deck: $deck,
-                direction: .up,
-                displayCount: 10,
-                alwaysShowLastCard: true,
-                verticalOffset: 10,
+                config: .standard,
                 swipeLeftAction: { hobby in print("\(hobby.id) was swiped left") },
                 swipeRightAction: { hobby in print("\(hobby.id) was swiped right") },
                 swipeUpAction: { hobby in print("\(hobby.id) was swiped up") },